Add locations tests for desugaring
